feat(tests): add deterministic timeout tests with conftest fixtures
Based on user guidance to fix timeout tests:
- Create conftest.py with tiny timeouts (0.05s) and FakeModel with async
- Create test_dual_streaming_timeouts.py with 5 timeout scenarios
- Change FLASH_TIMEOUT_S and PRO_TIMEOUT_S from int to float for sub-second precision
- Patch model _query methods to use async fake models (bypasses asyncio.to_thread)

Status: Tests run but SSE parsing needs debug (events list empty)
Next: Fix SSE parsing or client streaming setup

Tests to fix:
- test_normal_case_no_timeouts (sanity check)
- test_pro_timeout_emits_error_then_done
- test_flash_timeout_emits_error_then_done
- test_both_timeout_single_error_then_done
- test_stream_never_hangs_on_exception

+++ app/src/utils/settings.py
@@ -40,9 +40,9 @@ class Settings(BaseSettings):
     ENABLE_METRICS: bool = True
     ENABLE_TRACING: bool = True
     
-    # Dual-model timeout configuration (seconds)
-    FLASH_TIMEOUT_S: int = 5     # Flash target: <2s, allow 2.5x headroom
-    PRO_TIMEOUT_S: int = 45      # Pro target: ~20s, allow 2.25x headroom
+    # Dual-model timeout configuration (seconds - float for sub-second precision in tests)
+    FLASH_TIMEOUT_S: float = 5.0     # Flash target: <2s, allow 2.5x headroom
+    PRO_TIMEOUT_S: float = 45.0      # Pro target: ~20s, allow 2.25x headroom
     
     # Security settings
     API_KEY: Optional[str] = None  # API key for authentication (set in Secret Manager)
